work multiple grademe

+++ grademe.sh
@@ -6,44 +6,68 @@
 #    By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2019/10/29 21:24:29 by lmartin           #+#    #+#              #
-#    Updated: 2019/10/30 02:30:03 by lmartin          ###   ########.fr        #
+#    Updated: 2019/10/30 03:07:04 by lmartin          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 #! /bin/bash
 #!/usr/bin/osascript
 
-gcc -Wall -Wextra -Werror -c ../ft_printf.c > /dev/null 2>&1
-gcc -c main.c mine_system_d.c mine_system_s.c mine_system_x.c > /dev/null 2>&1
-gcc -o test_printf ft_printf.o main.o mine_system_d.o mine_system_s.o mine_system_x.o > /dev/null 2>&1
+compil ()
+{
+	gcc -Wall -Wextra -Werror -c ../ft_printf.c > /dev/null 2>&1
+	gcc -c main.c mine_system_d.c mine_system_s.c mine_system_x.c > /dev/null 2>&1
+	gcc -o test_printf ft_printf.o main.o mine_system_d.o mine_system_s.o mine_system_x.o > /dev/null 2>&1
+}
+
+start ()
+{
+	printf "============================================================\n"
+	printf "\e[92mType 'grademe' to be graded, good luck.\e[0m\n"
+	while true; do
+	    read input
+	    if [[ $input = "grademe" ]]; then
+	        grademe
+			break
+	    fi
+	done
+}
 
 tester () {
 	./test_printf "$1" > this_out
 	#TIMING=$( (time ./test_printf "$1" > this_out_$1) | grep "real" | cut -f 2)
-	if [ -n "$(cmp --silent "$1" "this_out")" ]
+	if [ "$(diff "$1" "this_out")" ]
 	then
-  	printf "$1 \e[31m[FAILED]\e[0m\n"
+		rm "this_out"
+  		printf "$1 \e[31m[FAILED]\e[0m\n"
+		start
 	else
-  	printf "$1 \e[92m[OK]\e[0m\n"
+		rm "this_out"
+  		printf "$1 \e[92m[OK]\e[0m\n"
 	fi
 	#echo "done testing \"$1\": "
 	#printf "[OK] %-25s (%s)\n" "\"$1\"" "$TIMING"
 }
 
 grademe() {
+	compil
+	if [ ! -f test_printf ]; then
+		printf "\e[31m[COMPILATION FAILED]\e[0m\n"
+		printf "============================================================\n"
+		exit
+	fi
 	printf "============================================================\n"
 	touch "this_out"
 	tester "system_d"
 	tester "system_s"
 	tester "system_x"
-	echo "============================================================"
-	rm "this_out"
+ 	printf "============================================================\n"
 	rm "main.o"
 	rm "mine_system_d.o"
 	rm "mine_system_s.o"
 	rm "mine_system_x.o"
 	rm "ft_printf.o"
-	echo "done testing"
 	rm "test_printf"
+	exit
 }
 clear
 printf "\e[34m                               \e[0m\n"
@@ -59,11 +83,6 @@ printf "\e[100m            ~ TRAINING PRINTF FOR EXAM RANK 02 ~            \e[0m
 printf "Your file must be in ../ to be corrected, and named ft_print\n.c otherwise, you will get \e[31m[COMPILATION FAILED]\e[0m flag\n"
 printf "============================================================\n"
 
-if [ ! -f test_printf ]; then
-	printf "\e[31m[COMPILATION FAILED]\e[0m\n"
-	printf "============================================================\n"
-	exit
-fi
 printf "If you want to start the test, type '\e[92myes\e[0m', a 120mn countdown\n will start.\n"
 read yes
 printf "============================================================\n"
@@ -73,37 +92,4 @@ if [ "$yes" != "yes" ]; then
 fi
 printf "\e[35msubject.en.txt (and fr) is available in /subjects\e[0m\n\n"
 cat	subjects/subject.en.txt
-printf "============================================================\n"
-hour=2
-min=0
-sec=0
-while [ $hour -ge 0 ]; do
-	while [ $min -ge 0 ]; do
-		while [ $sec -ge 0 ]; do
-			timer="0$hour:"
-			if ((min < 10)); then
-				timer="${timer}0$min:"
-			else
-				timer="${timer}$min:"
-			fi
-			if ((sec < 10)); then
-				timer="${timer}0$sec"
-			else
-				timer="${timer}$sec"
-			fi
-			printf "\r\e[33m[\e[0m\e[92m$timer\e[0m\e[33m]\e[0m>"
-			let "sec=sec-1"
-			read -t 1 -n 1 input
-		    if [[ $input = "q" ]] || [[ $input = "Q" ]]; then
-				# The following line is for the prompt to appear on a new line.
-		        clear
-		        exit
-		    fi
-		done
-		sec=59
-		let "min=min-1"
-	done
-	min=59
-	let "hour=hour-1"
-done
-printf "DONE"
+start
